% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_dup}
\alias{check_dup}
\title{Check Duplicates}
\usage{
check_dup(.source, .target, .check = c("source", "all"))
}
\arguments{
\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Target Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.check}{Check only column that are also in source, or all columns}
}
\value{
A list with duplicates
}
\description{
Description
}
