% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_dups}
\alias{check_dups}
\title{Helper Function: Check if column contains duplicates}
\usage{
check_dups(.tab, .col)
}
\arguments{
\item{.tab}{A dataframe}

\item{.col}{A column as string}
}
\value{
Nothing or Error
}
\description{
Helper Function: Check if column contains duplicates
}
