% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_id}
\alias{check_id}
\title{Check ID COlumns}
\usage{
check_id(.source, .target, .error = TRUE)
}
\arguments{
\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Target Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.error}{Return Error?}
}
\value{
Either Errors or a list
}
\description{
Description
}
