% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_matching_cols}
\alias{check_matching_cols}
\title{Helper Function: Check if all matching columns are present}
\usage{
check_matching_cols(.path, .cols, .type = c("s", "t"))
}
\arguments{
\item{.path}{A path to the stored tables}

\item{.cols}{matching columns}

\item{.type}{Either s or t}
}
\value{
Nothing or Error
}
\description{
Helper Function: Check if all matching columns are present
}
