% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_names}
\alias{check_names}
\title{Helper Function: Check for named vector}
\usage{
check_names(.tab, .cols)
}
\arguments{
\item{.tab}{A dataframe (either the source or target dataframe)}

\item{.cols}{A named character, with the columns names as string you want to match.\cr
The vector must be named wit either fuzzy (f) of exact (e).}
}
\value{
Nothing or an error
}
\description{
Helper Function: Check for named vector
}
