% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_nas}
\alias{check_nas}
\title{Check columns for NA values}
\usage{
check_nas(.source, .target, .check = c("source", "all"))
}
\arguments{
\item{.source}{The Source Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.target}{The Target Dataframe.
Must contain a unique column id and the columns you want to match on}

\item{.check}{Check only column that are also in source, or all columns}
}
\value{
A list with the number of NAs
}
\description{
Description
}
