% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_split_input}
\alias{check_split_input}
\title{Split and Check Input}
\usage{
check_split_input(.source, .target, .cols_match, .mat_size = 1000)
}
\arguments{
\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}

\item{.mat_size}{Maximal Matrix Size}
}
\value{
A dataframe or an error
}
\description{
Split and Check Input
}
