% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_dups}
\alias{filter_dups}
\title{Check for Duplicates}
\usage{
filter_dups(.tab, ...)
}
\arguments{
\item{.tab}{A dataframe}

\item{...}{Any number of columns of the dataframe}
}
\value{
A dataframe
}
\description{
Check for Duplicates
}
