% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_groups}
\alias{filter_groups}
\title{Filter data from fst files}
\usage{
filter_groups(.dir, .row)
}
\arguments{
\item{.dir}{Directory to store Tables}

\item{.row}{One row corrsponds to a group}
}
\value{
Filtered Dataframe
}
\description{
Filter data from fst files
}
