% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_datasets.R
\docType{data}
\name{legal_form_all}
\alias{legal_form_all}
\title{Dataset: Combined Legal Forms from the GLEIF and ECB}
\format{
A data frame with 4,933 rows and 3 variables:
\describe{
\item{id}{Legal Form ID}
\item{country_code}{ISO 3166-1 alpha-3 country codes}
\item{lfo}{Original legal form}
\item{lfs}{Standardized legal form}
\item{source}{Source}
\item{ids}{Original IDs}
}
}
\usage{
legal_form_all
}
\description{
Dataset: Combined Legal Forms from the GLEIF and ECB
}
\keyword{datasets}
