% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_datasets.R
\docType{data}
\name{legal_form_gle}
\alias{legal_form_gle}
\title{Dataset: Legal Forms from the GLEIF}
\format{
A data frame with 3,322 rows and 3 variables:
\describe{
\item{id}{Legal Form ID}
\item{iso3}{ISO 3166-1 alpha-3 country codes}
\item{lfo}{Original legal form}
\item{lfs}{Standardized legal form}
}
}
\usage{
legal_form_gle
}
\description{
Dataset: Legal Forms from the GLEIF
}
\keyword{datasets}
