% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_groups_vec}
\alias{make_groups_vec}
\title{Make Groups}
\usage{
make_groups_vec(.vec, .max)
}
\arguments{
\item{.vec}{A Numeric Vector}

\item{.max}{Maximal groups size}
}
\value{
A Numeric Vectore
}
\description{
Make Groups
}
