% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{match_group}
\alias{match_group}
\title{Helper Function: Match a single Group}
\usage{
match_group(
  .lst,
  .max_match = 10,
  .method = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex")
)
}
\arguments{
\item{.lst}{A list produced by filter_groups()}

\item{.max_match}{Maximum number of matches}

\item{.method}{c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex")}
}
\value{
A dataframe
}
\description{
Helper Function: Match a single Group
}
