% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{msg_verbose}
\alias{msg_verbose}
\title{Helper Function: Print Messages}
\usage{
msg_verbose(.msg, .verbose)
}
\arguments{
\item{.msg}{A character string}

\item{.verbose}{TRUE/FALSE}
}
\value{
A message
}
\description{
Helper Function: Print Messages
}
