% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_table}
\alias{prep_table}
\title{Helper Function: Prepare Table}
\usage{
prep_table(.tab, .fstd = NULL, .dir, .type = c("s", "t"))
}
\arguments{
\item{.tab}{A dataframe (either the source or target dataframe)}

\item{.fstd}{Standardization Function}

\item{.dir}{Directory to store Tables}

\item{.type}{Either s (source) or t (target)}
}
\value{
A dataframe
}
\description{
Description
}
