% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{split_input}
\alias{split_input}
\title{Split Inputs}
\usage{
split_input(
  .tab,
  .cols_exact = character(),
  .split = Inf,
  .type = c("source", "target")
)
}
\arguments{
\item{.tab}{Either Source or Target Dataframe}

\item{.cols_exact}{Columns used for Splitting}

\item{.split}{Maximum Number of Items to process in the Source Dataframe}

\item{.type}{c("source", "target")}
}
\value{
A nested Dataframe
}
\description{
Split Inputs
}
