% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{test_memory}
\alias{test_memory}
\title{Match Data}
\usage{
test_memory(
  .powers = 5,
  .source,
  .target,
  .cols_match,
  .max_match = 10,
  .method = "osa",
  .workers = future::availableCores()
)
}
\arguments{
\item{.powers}{Powers of 2 to test the Memory (default = 5 => 1,2,4,8,16,32)}

\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}

\item{.max_match}{Maximum number of matches to return (Default = 10)}

\item{.method}{One of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex".\cr
See: stringdist-metrics {stringdist}}

\item{.workers}{Number of cores to utilize (Default all cores determined by future::availableCores())}
}
\value{
A Dataframe
}
\description{
Match Data
}
