% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{uniqueness_data}
\alias{uniqueness_data}
\title{Score Data}
\usage{
uniqueness_data(.matches, .source, .target, .cols_match)
}
\arguments{
\item{.matches}{Dataframe produced by match_data()}

\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}
}
\value{
A dataframe
}
\description{
Description
}
