% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_id}
\alias{check_id}
\title{Helper Function: Check if table has a column called 'id'}
\usage{
check_id(.tab, .type = c("s", "t"))
}
\arguments{
\item{.tab}{A dataframe}

\item{.type}{Either s or t}
}
\value{
Nothing or Error
}
\description{
Helper Function: Check if table has a column called 'id'
}
