% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_names}
\alias{check_names}
\title{Helper Function: Check if matching vector is names}
\usage{
check_names(.cols)
}
\arguments{
\item{.cols}{columns in a vectore}
}
\value{
Nothing or Error
}
\description{
Helper Function: Check if matching vector is names
}
