% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_groups}
\alias{filter_groups}
\title{Helper Function: Filter data from fst files}
\usage{
filter_groups(.dir, .group)
}
\arguments{
\item{.dir}{Directory to store Tables}

\item{.group}{One row corrsponds to a group}
}
\value{
Filtered Dataframe
}
\description{
Helper Function: Filter data from fst files
}
