% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_groups}
\alias{make_groups}
\title{Helper Function: Make Groups}
\usage{
make_groups(.dir, .range = Inf)
}
\arguments{
\item{.dir}{Directory to store Tables}

\item{.range}{Character range}
}
\value{
A table with Groups (saved in .dir)
}
\description{
Helper Function: Make Groups
}
