% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{match_data}
\alias{match_data}
\title{Match Data}
\usage{
match_data(
  .source,
  .target,
  .cols_match,
  .max_match = 10,
  .method = "osa",
  .workers = future::availableCores(),
  .split = Inf,
  .verbose = TRUE
)
}
\arguments{
\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}

\item{.max_match}{Maximum number of matches to return (Default = 10)}

\item{.method}{One of "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex".\cr
See: stringdist-metrics {stringdist}}

\item{.workers}{Number of cores to utilize (Default all cores determined by future::availableCores())}

\item{.split}{Maximum Number of Items to process in the Source Dataframe}

\item{.verbose}{Print Additional Information?}
}
\value{
A Dataframe
}
\description{
Match Data
}
