% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{match_data}
\alias{match_data}
\title{Match Data}
\usage{
match_data(
  .dir,
  .max_match = 10,
  .method = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex"),
  .workers = floor(future::availableCores()/4),
  .verbose = TRUE
)
}
\arguments{
\item{.dir}{Directory to store Tables}

\item{.max_match}{MAximum number of matches}

\item{.method}{c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex")}

\item{.workers}{Workers to use}

\item{.verbose}{TRUE/FALSE}
}
\value{
A dataframe (saved in .dir)
}
\description{
Match Data
}
