% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{match_data}
\alias{match_data}
\title{Match Data}
\usage{
match_data(
  .dir,
  .cols,
  .range = Inf,
  .weights = NULL,
  .max_match = 10,
  .allow_mult = TRUE,
  .method = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex"),
  .workers = floor(future::availableCores()/4),
  .mat_size = 1e+07,
  .verbose = TRUE
)
}
\arguments{
\item{.dir}{Directory to store Tables}

\item{.cols}{A named character, with the columns names as string you want to match.\cr
The vector must be named wit either fuzzy (f) of exact (e).}

\item{.range}{A range of characters as an integer, e.g. the name of the source dataframe is 5 characters long and the
.range argument is 3, then all names in the target dataframe between 2 - 8 characters are used for the matching}

\item{.weights}{A named vector with weight, if null all columns will be equal weighted}

\item{.max_match}{Maximum number of matches}

\item{.allow_mult}{Aloow multiple matches?}

\item{.method}{c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex")}

\item{.workers}{Workers to use}

\item{.mat_size}{Maximum Matrix Size for Fuzzy Matching}

\item{.verbose}{Print additional Information? (Default: TRUE)}
}
\value{
A dataframe (saved in .dir)
}
\description{
Match Data
}
