% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_tables}
\alias{prep_tables}
\title{Prepare Tables for Matching}
\usage{
prep_tables(
  .source,
  .target,
  .cols,
  .fstd = standardize_str,
  .dir,
  .range = Inf
)
}
\arguments{
\item{.source}{Source Dataframe}

\item{.target}{Target Dataframe}

\item{.cols}{A named character, with the columns names as string you want to match.\cr
The vector must be named wit either fuzzy (f) of exact (e).}

\item{.fstd}{Standardization Function}

\item{.dir}{Directory to store Tables}

\item{.range}{Character range}
}
\value{
Dataframes (saved in .dir)
}
\description{
Prepare Tables for Matching
}
