% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{prep_tables}
\alias{prep_tables}
\title{Prepare Table}
\usage{
prep_tables(.tab, .cols_match)
}
\arguments{
\item{.tab}{A dataframe (either the source or target dataframe)}

\item{.cols_match}{A character vector of columns to perform fuzzy matching.}
}
\value{
A dataframe
}
\description{
Description
}
