% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_tables}
\alias{prep_tables}
\title{Prepare Tables for Matching}
\usage{
prep_tables(
  .source,
  .target,
  .fstd = NULL,
  .dir,
  .return = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{.source}{Source Dataframe}

\item{.target}{Target Dataframe}

\item{.fstd}{Standardization Function (Default: standardize_str())}

\item{.dir}{Directory to store Tables}

\item{.return}{Return a saved file(s)?}

\item{.verbose}{Print additional Information? (Default: TRUE)}
}
\value{
Dataframes (saved in .dir)
}
\description{
Prepare Tables for Matching
}
