% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{score_data}
\alias{score_data}
\title{Score Data}
\usage{
score_data(
  .dir,
  .weights = NULL,
  .max_match = 10,
  .method = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex")
)
}
\arguments{
\item{.dir}{Directory to store Tables}

\item{.weights}{Weights}

\item{.max_match}{Maximum number of matches}

\item{.method}{c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", "soundex")}
}
\value{
A dataframe (saved in .dir)
}
\description{
Score Data
}
