% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{score_data}
\alias{score_data}
\title{Title}
\usage{
score_data(.matches, .uniqueness = NULL, .weights = NULL)
}
\arguments{
\item{.matches}{Dataframe generated by match_data()}

\item{.uniqueness}{Dataframe generated by uniqueness_data()}

\item{.weights}{Named numeric vectore (Names must correspond to the similarity columns in .matches)}
}
\value{
A Dataframe
}
\description{
Title
}
