% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{select_data}
\alias{select_data}
\title{Select Data}
\usage{
select_data(
  .scores,
  .source,
  .target,
  .mult_match = FALSE,
  .cols_get = character()
)
}
\arguments{
\item{.scores}{Dataframe generated by match_data()}

\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.mult_match}{Allow for multile matches in the Target Dataframe}

\item{.cols_get}{Column to join}
}
\value{
A Dataframe
}
\description{
Select Data
}
