% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{select_data}
\alias{select_data}
\title{Select Data}
\usage{
select_data(.matches, .uniqueness = NULL, .source, .target, .weights = NULL)
}
\arguments{
\item{.matches}{Dataframe generated by match_data()}

\item{.uniqueness}{Dataframe generated by uniqueness_data()}

\item{.source}{The Source Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.target}{The Target Dataframe.\cr
(Must contain a unique column id and the columns you want to match on)}

\item{.weights}{Named numeric vectore (Names must correspond to the similarity columns in .matches)}
}
\value{
A Dataframe
}
\description{
Select Data
}
