% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{standardize_str}
\alias{standardize_str}
\title{Standardize Strings}
\usage{
standardize_str(.str, .op = c("space", "punct", "case", "ascii"))
}
\arguments{
\item{.str}{A character vector}

\item{.op}{Any of c("space", "punct", "case", "ascii")}
}
\value{
A string
}
\description{
Description
}
\examples{
standardize_str(c("jkldsa   jkdhas   äää  §$ ## #'''"))
standardize_str(c("jkldsa   jkdhas   äää  §$ ## #'''"), "space")
}
